Update chained_CB3_II.m
Fixed two typos in the function "Chained CB3 II" (test problem 5). These typos caused the objective values to be wrong, causing the convergence issues observed for Hanso and Granso in the article. The arXiv version of the article was updated with the fixed results. (This does not impact our interpretation of the results.)